calibration only goes down
